# docs/api/commerceflow-openapi.yaml
...
openapi: 3.0.3
info:
  title: CommerceFlow API
  version: 1.0.0
  description: |
    Microservices-based e-commerce order processing.

    Placing an order is **asynchronous**. `POST /api/orders` returns `201` with status `CREATED`
    as soon as the order is accepted and the orchestrator has been started; stock
    reservation, payment and notification then run over Kafka. Poll `GET /api/orders/{id}` to
    follow it — the status walks
    `CREATED → INVENTORY_RESERVED → PAID → COMPLETED`, or moves to `CANCELLED` with
    `failedStep` and `failureReason` telling you where and why it stopped.

    Every response is wrapped in a uniform envelope: `ApiResponse` on success, `ErrorResponse`
    on failure. Both carry the `correlationId`, which is the id to quote in a support request —
    it stitches the whole saga together in the logs.
  contact:
    name: CommerceFlow Platform Team
    email: platform@commerceflow.io
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080
    description: Local API Gateway (docker compose)
  - url: https://api.commerceflow.io
    description: Production
...
security:
  - bearerAuth: []
paths:
...
  /api/orders/{orderId}/returnable:
    get:
      tags: [Returns]
      summary: What of this order can still be sent back
      description: >-
        Every line with what was ordered, what an open or settled return has already claimed, and
        what one unit gives back. Served before the customer chooses, so a form cannot offer a
        quantity that would be refused.
      parameters:
        - { name: orderId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: One entry per order line
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ReturnableLine" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/orders/{orderId}/returns:
    get:
      tags: [Returns]
      summary: Returns raised against one order
      parameters:
        - { name: orderId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Newest first
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Return" }
    post:
      tags: [Returns]
      summary: Ask to send items back
      description: |
        Only for an order that has **COMPLETED**. One still being paid for is cancelled instead —
        that unwinds the saga and puts the stock straight back, which is a different and cheaper
        operation.

        The refund is worked out and frozen now, from the prices actually charged: a line gives
        back its `unitPrice` (which already has any coupon allocated into it) plus tax at the rate
        frozen on the order. A catalogue change or an expiring campaign between asking and
        refunding cannot alter what somebody gets.

        Delivery is refunded only when the whole order comes back. A customer keeping one item of
        three has still had the parcel delivered.

        Refused with `409` when a line is asked for in a quantity larger than what is left after
        earlier returns.
      parameters:
        - { name: orderId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateReturnRequest" }
      responses:
        "201":
          description: The return, priced
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Return" }
        "409": { $ref: "#/components/responses/Conflict" }
...
  /api/returns/{returnId}:
    get:
      tags: [Returns]
      summary: One return
      description: A customer sees their own; somebody else's returns 404 rather than 403.
      parameters:
        - { name: returnId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: The return
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Return" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/returns/{returnId}/cancel:
    post:
      tags: [Returns]
      summary: Call off a return
      description: >-
        Only while nobody has decided on it. Once it is approved the goods are on their way back,
        and cancelling would leave a parcel arriving that nothing expects.
      parameters:
        - { name: returnId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Cancelled
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Return" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/returns:
    get:
      tags: [Returns]
      summary: The returns queue
      description: >-
        Administrator only. Filtered by state, oldest first — which is the order it should be
        worked in. Without a state, everything newest first.
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [REQUESTED, APPROVED, REJECTED, RECEIVED, REFUND_PENDING, REFUNDED, CANCELLED]
        - { name: page, in: query, schema: { type: integer, default: 0, minimum: 0 } }
        - { name: size, in: query, schema: { type: integer, default: 20, minimum: 1, maximum: 100 } }
      responses:
        "200": { description: A page of returns }
        "403": { description: Not an administrator }
  /api/returns/{returnId}/approve:
    post:
      tags: [Returns]
      summary: Approve a return
      description: >-
        Administrator only. The customer is told to send the goods back. Nothing is refunded yet —
        money follows the goods, not the promise of them.
      parameters:
        - { name: returnId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ReturnDecision" }
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Return" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/returns/{returnId}/reject:
    post:
      tags: [Returns]
      summary: Refuse a return
      description: >-
        Administrator only. The note is shown to the customer and is the only place they will find
        out why. A refusal with no reason produces a support ticket instead.
      parameters:
        - { name: returnId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ReturnDecision" }
      responses:
        "200":
          description: Refused
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Return" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/returns/{returnId}/received:
    post:
      tags: [Returns]
      summary: The goods are back
      description: |
        Administrator only. Separate from refunding on purpose: what came back is a warehouse fact
        and what is paid is a money decision, and one person often does the first without the
        authority for the second.

        **`restock` is required and has no default.** Not everything that comes back can be sold
        again — a change of mind returns to the shelf, a cracked screen does not — and only the
        person holding the box can tell. `true` sends a restock command to Inventory Service for
        exactly the returned lines, back into the building they shipped from; `false` records the
        units as received and written off.
      parameters:
        - { name: returnId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [restock]
              properties:
                restock: { type: boolean }
      responses:
        "200":
          description: Recorded
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Return" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/returns/{returnId}/refund:
    post:
      tags: [Returns]
      summary: Send the money back
      description: |
        Administrator only. Asks Payment Service over `return.refund.commands` and leaves the
        return **REFUND_PENDING** until the answer arrives on `return.refunded`. There is no
        synchronous call: the command goes into the outbox in the same transaction as the state
        change, so the two cannot disagree.

        Pressing it again while pending does nothing, so two people working the same queue cannot
        refund twice. Payment Service also treats the return id as its idempotency key, which is a
        unique constraint there — a republished command refunds once.

        A failure puts the return back to **RECEIVED** with the provider's reason on it, so it can
        be retried once whatever went wrong is fixed.
      parameters:
        - { name: returnId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: The refund has been asked for
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Return" }
        "409": { $ref: "#/components/responses/Conflict" }

  # ------------------------------------------------------------------ reviews
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token returned by `POST /api/auth/login`

  # This file is the hand-written contract for the paths that matter most. It is deliberately
  # not exhaustive: each service serves its own generated definition at /v3/api-docs, aggregated
  # behind the gateway at /swagger-ui.html, and that is authoritative for everything else --
  # notably the back-office endpoints for warehouses, stock levels, product images and review
  # moderation, which no external client calls.
  responses:
...
    NotFound:
      description: No such resource
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
    Conflict:
      description: |
        The request is well formed but the resource is not in a state that allows it
        (`ORDER_NOT_MODIFIABLE`, `CONFLICT`). The message says what to do about it.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  schemas:
...
    Return:
      type: object
      description: >-
        A request to send goods back. Its lifecycle is separate from the order saga: the order
        stays COMPLETED throughout.
      properties:
        id: { type: string, format: uuid }
        orderId: { type: string, format: uuid }
        orderNumber: { type: string }
        status:
          type: string
          enum: [REQUESTED, APPROVED, REJECTED, RECEIVED, REFUND_PENDING, REFUNDED, CANCELLED]
          description: >-
            REQUESTED -> APPROVED -> RECEIVED -> REFUND_PENDING -> REFUNDED is the ordinary path.
            REJECTED ends it after a decision, CANCELLED when the customer calls it off before
            one. A failed refund returns it to RECEIVED with refundFailure set.
        reason:
          type: string
          description: The customer's own words, kept verbatim.
        refundAmount:
          type: number
          description: Frozen when the request is raised, from the prices actually charged.
        currency: { type: string, example: EUR }
        refundShipping:
          type: boolean
          description: True only when the whole order came back.
        items:
          type: array
          items: { $ref: "#/components/schemas/ReturnItem" }
        requestedAt: { type: string, format: date-time }
        decidedAt: { type: string, format: date-time, nullable: true }
        decisionNote:
          type: string
          nullable: true
          description: Shown to the customer. On a refusal it is the only explanation they get.
        receivedAt: { type: string, format: date-time, nullable: true }
        restocked:
          type: boolean
          nullable: true
          description: >-
            Whether the goods went back on sale. Null until they arrive — null and false are
            different answers: "not opened yet" against "opened and written off".
        refundedAt: { type: string, format: date-time, nullable: true }
        refundReference:
          type: string
          nullable: true
          description: The acquirer's reference, once there is one. What reconciliation matches on.
        refundFailure:
          type: string
          nullable: true
          description: Why the last refund attempt failed. The return can be retried.
    ReturnItem:
      type: object
      properties:
        orderItemId: { type: string, format: uuid }
        productId: { type: string, format: uuid }
        productName: { type: string }
        sku: { type: string }
        quantity: { type: integer }
        refundAmount:
          type: number
          description: >-
            Net at the price actually charged, plus tax at the rate frozen on the order. Computed
            per line and rounded once, never apportioned out of an order total.
    ReturnableLine:
      type: object
      properties:
        orderItemId: { type: string, format: uuid }
        productId: { type: string, format: uuid }
        productName: { type: string }
        sku: { type: string }
        orderedQuantity: { type: integer }
        alreadyReturned:
          type: integer
          description: >-
            Claimed by an open or settled return. A rejected or cancelled one does not count,
            because no goods moved.
        returnableQuantity: { type: integer }
        refundPerUnit: { type: number }
        currency: { type: string }
    CreateReturnRequest:
      type: object
      required: [items, reason]
      properties:
        items:
          type: array
          minItems: 1
          items:
            type: object
            required: [orderItemId, quantity]
            properties:
              orderItemId:
                type: string
                format: uuid
                description: >-
                  The order line, not the product: one order can carry the same product on two
                  lines and they may be returned separately.
              quantity: { type: integer, minimum: 1 }
        reason:
          type: string
          maxLength: 500
          description: >-
            Why, in the customer's own words. A faulty item and a change of mind need different
            handling, and a dropdown always lacks the case in front of you.
    ReturnDecision:
      type: object
      properties:
        note:
          type: string
          maxLength: 500
          description: Shown to the customer.
...
    ErrorResponse:
      type: object
      description: Uniform error envelope.
      properties:
        success: { type: boolean, example: false }
        timestamp: { type: string, format: date-time }
        status: { type: integer, example: 409 }
        error: { type: string, example: Conflict }
        code:
          type: string
          example: INSUFFICIENT_STOCK
          description: Machine readable error code; stable across releases.
        message: { type: string }
        path: { type: string, example: /api/orders }
        correlationId: { type: string, nullable: true }
        fieldErrors:
          type: array
          nullable: true
          items:
            type: object
            properties:
              field: { type: string, example: email }
              rejectedValue: {}
              message: { type: string, example: must be a valid address }
...